Consolidate Pytorch train module and build_image.sh script
