+++ pytorch_mnist/training/ddp/mnist/build_image.sh
