+++ pytorch_mnist/training/ddp/mnist/cpu/Dockerfile
@@ -38,5 +38,5 @@ RUN pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org to
 WORKDIR /workspace
 RUN chmod -R a+w /workspace
 
-ADD . /opt/pytorch_dist_mnist
-ENTRYPOINT ["mpirun", "-n", "4", "--allow-run-as-root", "python", "-u", "/opt/pytorch_dist_mnist/mnist_DDP_CPU.py"]
+ADD ../mnist_DDP.py /opt/pytorch_dist_mnist
+ENTRYPOINT ["mpirun", "-n", "4", "--allow-run-as-root", "python", "-u", "/opt/pytorch_dist_mnist/mnist_DDP.py", "/mnt/kubeflow-gcfs/pytorch/model"]
